fix: loglevel wasn't respected
