+++ jcloud/flow.py
@@ -12,41 +12,15 @@ from typing import Dict, List, Optional
 import aiohttp
 from rich import print
 
+from .constants import WOLF_API, LOGSTREAM_API, ARTIFACT_API, Status
 from .helper import get_logger, get_or_reuse_loop, get_pbar, normalized, zipdir
 
-WOLF_API = 'https://api.wolf.jina.ai/dev/flows'
-LOGSTREAM_API = 'wss://logs.wolf.jina.ai/dev'
-ARTIFACT_API = 'https://apihubble.jina.ai/v2/rpc/artifact.upload'
 
 logger = get_logger()
 
 pbar, pb_task = get_pbar(
     '', total=5, disable='JCLOUD_NO_PROGRESSBAR' in os.environ
 )  # progress bar for deployment
-
-
-class Status(str, Enum):
-    SUBMITTED = 'SUBMITTED'
-    NORMALIZING = 'NORMALIZING'
-    NORMALIZED = 'NORMALIZED'
-    STARTING = 'STARTING'
-    FAILED = 'FAILED'
-    ALIVE = 'ALIVE'
-    UPDATING = 'UPDATING'
-    DELETING = 'DELETING'
-    DELETED = 'DELETED'
-
-    @property
-    def streamable(self) -> bool:
-        return self in (Status.ALIVE, Status.UPDATING, Status.DELETING)
-
-    @property
-    def alive(self) -> bool:
-        return self == Status.ALIVE
-
-    @property
-    def deleted(self) -> bool:
-        return self == Status.DELETED
 
 
 def _exit_if_response_error(response, expected_status):
